// frontend/web-app/src/app/api/model/semesterVO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SubjectDTO } from './subjectDTO';


export interface SemesterVO {
    degreeIds?: Array<number>;
    fromDate?: Date;
    semesterType?: string;
    subjects?: Array<SubjectDTO>;
    toDate?: Date;
}